refactor: change name to automatic-compilation
