+++ src/app/logic/automatic-compilation.ts
