+++ src/app/pages/LandingPage.svelte
@@ -5,7 +5,7 @@
   import LanguageNameInput from "../components/LanguageNameInput.svelte";
   import SplitButton from "../components/SplitButton.svelte";
   import { currentDownloadURL } from "../stores";
-  import { setupAutomaticCompilationAndDownloadURL } from "../logic/im-not-sure-yet";
+  import { setupAutomaticCompilationAndDownloadURL } from "../logic/automatic-compilation";
 
   let languageStatus: boolean = false;
   let continueReady: boolean = false;
